Fix Compiler Warnings

+++ Source/Core/Renderer/ERS_CLASS_VisualRenderer/ERS_CLASS_VisualRenderer.cpp
@@ -329,7 +329,7 @@ void ERS_CLASS_VisualRenderer::UpdateViewport(int Index, ERS_CLASS_SceneManager*
         
             long ScriptIndex = Model->AttachedScriptIndexes_[x];
         
-            if (ScriptIndex > -1 && ScriptIndex < ProjectUtils_->ProjectManager_->Project_.Scripts.size()) {
+            if (ScriptIndex > -1 && ScriptIndex < (long)ProjectUtils_->ProjectManager_->Project_.Scripts.size()) {
                 NewIndexes.push_back(ScriptIndex);
             }
